+++ src/app/(cn)/(homepage)/page.mdx
@@ -3,9 +3,9 @@ import { Resources } from '@/components/Resources'
 import { HeroPattern } from '@/components/HeroPattern'
 
 export const metadata = {
-  title: 'Protium',
+  title: 'SciOL',
   description:
-    'Learn everything there is to know about the Protium and integrate tools into your product.',
+    '共同迈向 SGI(Science General Intelligence)的未来',
   image: '/images/python.svg',
 }
 
@@ -16,7 +16,7 @@ export const sections = [
 
 <HeroPattern />
 
-# Protium Tutorial
+# ScienceOL Docs
 
 Use the Tutorial to access our features, knowledges, and seamlessly integrate your 
 research into the workflows of dozens of devoted Protium explorers. {{ className: 'lead' }}
