+++ src/app/(cn)/development/components/Resources.tsx
