+++ src/app/(cn)/development/components/index.ts
@@ -1,0 +1,2 @@
+export { Guides } from './Guides'
+export { Resources } from './Resources'
