+++ src/app/(cn)/development/deploys/tag-deploy/page.mdx
@@ -1,0 +1,132 @@
+
+
+# 🔄 main 分支上线流程
+
+### 触发事件：在 `dev` 分支创建标签 `v1.0.0`
+
+```bash
+git checkout dev
+git tag v1.0.0
+git push origin v1.0.0
+```
+
+或者使用一键部署脚本
+
+```bash
+./launch/tag.sh
+```
+
+### 详细执行流程：
+
+```mermaid
+graph TD
+    A[🏷️ 在 dev 分支创建标签 v1.0.0] --> B[推送标签到远程仓库]
+    
+    B --> C[🚀 触发 buildx-latest-on-tag.yaml]
+    B --> D[🔄 触发 merge-tag-to-main.yaml]
+    
+    C --> C1[验证标签来自 dev 分支]
+    C1 --> C2[构建 latest 镜像]
+    C2 --> C3[推送到阿里云镜像仓库]
+    C3 --> C4[创建 GitHub Release]
+    C4 --> C5[发送构建通知邮件]
+    
+    D --> D1[验证标签格式 SemVer]
+    D1 --> D2[获取标签指向的提交]
+    D2 --> D3[合并标签到 main 分支]
+    D3 --> D4[推送 main 分支]
+    D4 --> D5[发送合并通知邮件]
+    
+    D4 --> E[🔄 触发 sync-main-to-dev.yaml]
+    
+    E --> E1[检出 dev 分支]
+    E1 --> E2[获取最新 main 分支]
+    E2 --> E3[将 dev 基于 main 进行 rebase]
+    E3 --> E4[添加 [skip ci] 标记]
+    E4 --> E5[强制推送 dev 分支]
+    
+    E5 --> F[❌ buildx-when-push-dev.yaml 被跳过]
+    F --> G[✅ 流程完成，避免循环触发]
+    
+    style A fill:#e1f5fe
+    style C fill:#fff3e0
+    style D fill:#f3e5f5
+    style E fill:#e8f5e8
+    style F fill:#ffebee
+    style G fill:#e8f5e8
+```
+
+## 📋 详细步骤说明
+
+### 阶段 1：标签发布（并行执行）
+
+**🚀 buildx-latest-on-tag.yaml** - 构建生产镜像
+- ✅ 验证标签来自 `dev` 分支
+- 🐳 构建多架构 Docker 镜像（amd64）
+- 📤 推送到阿里云镜像仓库：`latest` 标签
+- 📝 使用 AI 生成 Release Notes
+- 🎉 创建 GitHub Release
+- 📧 发送构建成功通知
+
+**🔄 merge-tag-to-main.yaml** - 合并到主分支
+- ✅ 验证标签符合 SemVer 格式（如 `v1.0.0`）
+- 📍 获取标签指向的具体提交 SHA
+- 🔀 将该提交合并到 `main` 分支（使用 `--no-ff`）
+- 📤 推送更新后的 `main` 分支
+- 📧 发送合并通知邮件
+
+### 阶段 2：同步回开发分支
+
+**🔄 sync-main-to-dev.yaml** - 同步主分支到开发分支
+- 🎯 触发条件：`main` 分支有新推送 OR "Merge Tag to Main" 工作流完成
+- 📥 检出 `dev` 分支
+- 📤 获取最新的 `main` 分支
+- 🔄 将 `dev` 分支基于 `main` 进行 rebase
+- ⚠️ 冲突时自动选择 `dev` 的代码
+- 🏷️ **关键**：在提交消息中添加 `[skip ci]` 标记
+- 🚫 强制推送 `dev` 分支
+
+### 阶段 3：避免循环触发
+
+**❌ buildx-when-push-dev.yaml** - 开发环境构建（被跳过）
+- 🚫 由于提交消息包含 `[skip ci]`，GitHub Actions 自动跳过此工作流
+- ✅ 成功避免循环触发和不必要的构建
+
+## 🎯 关键改进点
+
+### 1. **避免循环触发** ✅
+- 使用 `[skip ci]` 标记防止 rebase 操作触发开发环境构建
+- 节省 CI 资源，避免无意义的构建
+
+### 2. **双重触发器** ✅
+```yaml
+on:
+  push:
+    branches:
+      - main
+  workflow_run:
+    workflows: ['Merge Tag to Main']
+    types:
+      - completed
+```
+
+### 3. **智能冲突处理** ✅
+```bash
+git rebase main || {
+  git diff --name-only --diff-filter=U | xargs -I {} git checkout --ours {}
+  git add -A
+  git rebase --continue
+}
+```
+
+## 📊 预期结果
+
+执行完整个流程后，您会得到：
+
+1. **🎉 GitHub Release**：包含 AI 生成的发布说明
+2. **🐳 生产镜像**：`registry.cn-shanghai.aliyuncs.com/scienceol/service:latest`
+3. **📧 邮件通知**：构建和合并状态通知
+4. **🔄 同步分支**：`main` 和 `dev` 保持同步
+5. **⚡ 高效执行**：避免不必要的 CI 构建
+
+这个设计既保证了版本发布的完整性，又维护了分支间的同步，同时避免了资源浪费！
