+++ src/app/(cn)/development/models/workflow/page.mdx
@@ -1,0 +1,5 @@
+export const metadata = {
+    title: "工作流模型",
+    description: "工作流数据模型"
+}
+
