+++ src/app/(cn)/development/page.mdx
@@ -1,9 +1,9 @@
-import { Guides } from './Guides'
-import { Resources } from './Resources'
+import { Guides } from './components'
+import { Resources } from './components'
 import { HeroPattern } from '@/components/HeroPattern'
 
 export const metadata = {
-  title: 'Protium 开发者中心',
+  title: '开发者中心',
   description:
     '为开发者提供完整的 Protium 开发指南、部署文档和文档编写规范。',
   image: '/images/python.svg',
